Add a solution for Day 1 Part 1.
A pretty straightforward warmup problem. Nothing special here except
the input parsing. After having done these sorts of things quite a
bit, I find getting a generic input parsing routine makes solving a
little bit easier. Basically it just splits and maps the input, but
this is generic enough that we should be able to use it for most
inputs.