Add a solution for Day 2 Part 1.
This problem had a lot of setup, but otherwise made sense. We just need to
find the outcomes for these games, get the point value for each game, and
then sum them up.

Additionally I updated the input module to add splitting on newlines. This
also needed me to add a filter to exclide the last empty line.