Add a solution for Day 3 Part 2.
For this one we do not need to split the rucksacks in half so that is
nice, but we do need to expand on our array intersection function.
Before it only supported 2 arrays, but we need to support three so that
we can compare all rucksacks and find the items that are the same in all
three.

So we extend it by doing an intersection on each intermediate result with
all passed in arrays. This makes it generic and expand to as many arrays
as we care to add.

We also need to take chunks of three rucksacks so I added a method to chunk
an array. Notice how we typed the return tuple manually because there is
not another way to do such a thing that I know of.