Add a solution for Day 4 Part 1.
I just went with the most obvious (to me)
solution of checking if the intersection of
the ranges matches the length of one of the
ranges, then we know it is a subset.

This would probably time out in a larger data
set, but it does not break here so that is
okay.

If we had a larger data set we would probably
want to not generate actual arrays and just
check with the starts and ends.