Add a solution for Day 4 Part 2.
Luckily my solution for part 1 is a good
setup for part 2. Same approach, but this
time we count if the intersection has any
matching elements.

Again, this would probably time out in a
larger data set, and we would probably just
compare using the starts and ends if we
wanted to solve a larger data set instead
of generating the arrays.