Use a virtual range instead of allocating actual values for an array.
We create a "virtual" array that knows how to act like a range of values.

It also will fall back to a real array if it needs to.

We do this because problems may have many, many items in a range.
Say we need a range from 1 to 1,000,000. We would need to use quite
a bit of memory just to create the array. This avoids that by
generating the needed elements on demand.